Ensure order of soap arguments based on function signature.
When an associative array is passed it does not use the key for argument order.
This means that the following code passes the last_name as first argument,
whilst in the signature this is the second parameter.

```
$arguments = [
    'last_name' => 'Lastname',
    'first_name' => 'Firstname',
];

$results = $client->getAdapter()->call('complianceSearchPersons', $arguments);
```

In this merge request we are leveraging the __getFunctions from the native php soap-client and parse those for argument order.
We order the arguments according to the signature before sending them to the parent __soapCall.
In addition we validate if the method exists and log any arguments that are not in the signature.